pci: The irq can be larger than 8 bits

+++ pkg/pci/pci.go
@@ -27,7 +27,7 @@ type PCI struct {
 
 	Latency   byte
 	IRQPin    byte
-	IRQLine   uint8
+	IRQLine   uint
 	Bridge    bool
 	FullPath  string
 	ExtraInfo []string
